+++ Assets/Resources/Prefabs/Tower/Tower1.prefab
@@ -178,13 +178,14 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 7647195984229369159}
-  - component: {fileID: 5953190201795358970}
   - component: {fileID: 5953190201795358971}
   - component: {fileID: 973573703778125118}
   - component: {fileID: 1636111063}
+  - component: {fileID: 2914295280108778092}
+  - component: {fileID: 4852740721596803879}
   m_Layer: 0
   m_Name: Tower1
-  m_TagString: Untagged
+  m_TagString: Tower
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -205,27 +206,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!50 &5953190201795358970
-Rigidbody2D:
-  serializedVersion: 4
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5012109670245564110}
-  m_BodyType: 0
-  m_Simulated: 1
-  m_UseFullKinematicContacts: 0
-  m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 0
-  m_AngularDrag: 0.05
-  m_GravityScale: 0
-  m_Material: {fileID: 0}
-  m_Interpolate: 0
-  m_SleepingMode: 1
-  m_CollisionDetection: 0
-  m_Constraints: 0
 --- !u!58 &5953190201795358971
 CircleCollider2D:
   m_ObjectHideFlags: 0
@@ -254,6 +234,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: be6f3c47371e85c479319fa8102db5f6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  CanAttack: 1
 --- !u!210 &1636111063
 SortingGroup:
   m_ObjectHideFlags: 0
@@ -265,6 +246,56 @@ SortingGroup:
   m_SortingLayerID: -2021860623
   m_SortingLayer: 1
   m_SortingOrder: 0
+--- !u!114 &2914295280108778092
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5012109670245564110}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e897565043d900844ab88d00d4b92a79, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  towerPrefab: {fileID: 0}
+  weapon:
+  - sprite: {fileID: 0}
+    damage: 0
+    rate: 0
+    range: 0
+    cost: 0
+  - sprite: {fileID: 0}
+    damage: 0
+    rate: 0
+    range: 0
+    cost: 0
+  - sprite: {fileID: 0}
+    damage: 0
+    rate: 0
+    range: 0
+    cost: 0
+--- !u!50 &4852740721596803879
+Rigidbody2D:
+  serializedVersion: 4
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5012109670245564110}
+  m_BodyType: 0
+  m_Simulated: 1
+  m_UseFullKinematicContacts: 0
+  m_UseAutoMass: 0
+  m_Mass: 1
+  m_LinearDrag: 0
+  m_AngularDrag: 0.05
+  m_GravityScale: 0
+  m_Material: {fileID: 0}
+  m_Interpolate: 0
+  m_SleepingMode: 1
+  m_CollisionDetection: 0
+  m_Constraints: 0
 --- !u!1 &5046374962057924131
 GameObject:
   m_ObjectHideFlags: 0
