+++ Assets/Resources/Prefabs/Tower/Tower2.prefab
@@ -301,7 +301,7 @@ GameObject:
   - component: {fileID: 1595224387}
   m_Layer: 0
   m_Name: Tower2
-  m_TagString: Untagged
+  m_TagString: Tower
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -371,6 +371,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6cb5e8d11cfe98d448c496c1eb8aade2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  CanAttack: 1
 --- !u!210 &1595224387
 SortingGroup:
   m_ObjectHideFlags: 0
