+++ Assets/Resources/Prefabs/Tower/Tower3.prefab
@@ -301,7 +301,7 @@ GameObject:
   - component: {fileID: 1950821731}
   m_Layer: 0
   m_Name: Tower3
-  m_TagString: Untagged
+  m_TagString: Tower
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -371,6 +371,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 041332db338425445adbb5d9c44301a5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  CanAttack: 1
 --- !u!210 &1950821731
 SortingGroup:
   m_ObjectHideFlags: 0
