+++ Assets/Resources/Prefabs/Tower/Tower4.prefab
@@ -99,7 +99,7 @@ GameObject:
   - component: {fileID: 1960193018}
   m_Layer: 0
   m_Name: Tower4
-  m_TagString: Untagged
+  m_TagString: Tower
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -169,6 +169,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b6fdf598ac567dd40973728179d1068f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  CanAttack: 1
 --- !u!210 &1960193018
 SortingGroup:
   m_ObjectHideFlags: 0
